+++ rust/python-bindings/src/passwords_impl.rs
@@ -2,7 +2,6 @@
 // Use of this source code is governed by the Apache License 2.0
 // that can be found in the LICENSE file.
 
-use pyo3::Bound;
 use pyo3::PyResult;
 #[cfg(any(feature = "cbc", feature = "sha512", feature = "simple-7"))]
 use pyo3::exceptions::PyRuntimeError;
@@ -10,11 +9,6 @@ use pyo3::exceptions::PyRuntimeError;
 use pyo3::exceptions::PyValueError;
 #[cfg(any(feature = "cbc", feature = "sha512", feature = "simple-7"))]
 use pyo3::pyfunction;
-use pyo3::types::PyModule;
-#[cfg(any(feature = "cbc", feature = "sha512", feature = "simple-7"))]
-use pyo3::types::PyModuleMethods as _;
-#[cfg(any(feature = "cbc", feature = "sha512", feature = "simple-7"))]
-use pyo3::wrap_pyfunction;
 
 #[cfg(feature = "sha512")]
 #[pyfunction]
@@ -79,27 +73,4 @@ pub(crate) fn simple_7_decrypt(data: &str) -> PyResult<String> {
         ::passwords::Simple7Error::InvalidUtf8(_) => PyRuntimeError::new_err(err.to_string()),
         _ => PyValueError::new_err(err.to_string()),
     })
-}
-
-#[cfg(any(feature = "cbc", feature = "sha512", feature = "simple-7"))]
-pub(crate) fn register(module: &Bound<'_, PyModule>) -> PyResult<()> {
-    #[cfg(feature = "sha512")]
-    module.add_function(wrap_pyfunction!(sha512_crypt, module)?)?;
-    #[cfg(feature = "cbc")]
-    {
-        module.add_function(wrap_pyfunction!(cbc_encrypt, module)?)?;
-        module.add_function(wrap_pyfunction!(cbc_decrypt, module)?)?;
-        module.add_function(wrap_pyfunction!(cbc_verify, module)?)?;
-    }
-    #[cfg(feature = "simple-7")]
-    {
-        module.add_function(wrap_pyfunction!(simple_7_encrypt, module)?)?;
-        module.add_function(wrap_pyfunction!(simple_7_decrypt, module)?)?;
-    }
-    Ok(())
-}
-
-#[cfg(not(any(feature = "cbc", feature = "sha512", feature = "simple-7")))]
-pub(crate) fn register(_module: &Bound<'_, PyModule>) -> PyResult<()> {
-    Ok(())
 }
